6.12: Move cachy-dev patch to misc
For the sake of not bloating the tree and redundancy, move cachy-dev to misc. The only difference between this patchset
and the main patchset are no more than 200 lines of code so this is a big improvement. This also enables users that want
to use cachy-dev in 6.11 to use it.

Signed-off-by: Eric Naim <dnaim@cachyos.org>