6.12: Add le9uo
This patch is to be applied on top of `0001-cachyos-base-all.patch` (or atleast `0005-cachy.patch`).
This adds le9uo functionality to the kernel and enables it by default. Disables the thrashing protection by MGLRU set by cachy
because it is unnecessary and redundant (also helps to prevent potential clashes between the two implementations).

Signed-off-by: Eric Naim <dnaim@cachyos.org>